flags should not be common
